Adds coverage_model.utils package
Adds a utils package to coverage_model.  Moves the following functions, previously in various places:
* prod
* create_guid
* is_valid_constraint
* fix_slice

Fixes imports
