+++ coverage_model/parameter_values.py
@@ -7,8 +7,9 @@
 @brief Abstract and concrete value objects for parameters
 """
 from ooi.logging import log
-from coverage_model.basic_types import AbstractBase, is_valid_constraint, InMemoryStorage, VariabilityEnum
+from coverage_model.basic_types import AbstractBase, InMemoryStorage, VariabilityEnum
 from coverage_model.numexpr_utils import is_well_formed_where, nest_wheres
+from coverage_model.utils import fix_slice, prod
 import numpy as np
 import numexpr as ne
 
@@ -16,10 +17,6 @@ def get_value_class(param_type, **kwargs):
     module = __import__(param_type._value_module, fromlist=[param_type._value_class])
     classobj = getattr(module, param_type._value_class)
     return classobj(parameter_type=param_type, **kwargs)
-
-def prod(lst):
-    import operator
-    return reduce(operator.mul, lst, 1)
 
 #=========================
 # Abstract Parameter Value Objects
@@ -39,56 +36,6 @@ class AbstractParameterValue(AbstractBase):
         self.parameter_type = parameter_type
         self.domain_set = domain_set
         self._storage = storage or InMemoryStorage(dtype=self.storage_encoding, fill_value=self.parameter_type.fill_value)
-
-    def _fix_slice(self, slice_):
-        # CBM: First swack - see this for more possible checks: http://code.google.com/p/netcdf4-python/source/browse/trunk/netCDF4_utils.py
-        if not is_valid_constraint(slice_):
-            raise SystemError('invalid constraint supplied: {0}'.format(slice_))
-
-        # First, ensure we're working with a list so we can make edits...
-        if not np.iterable(slice_):
-            slice_ = [slice_,]
-        elif not isinstance(slice_,list):
-            slice_ = list(slice_)
-
-        # Then make sure it's the correct rank
-        rank = len(self.shape)
-        slen = len(slice_)
-        if not slen == rank:
-            if slen > rank:
-                slice_ = slice_[:rank]
-            else:
-                for n in xrange(slen, rank):
-                    slice_.append(slice(None,None,None))
-
-        # Next, deal with negative indices
-        # Logic for handling negative indices from: http://docs.scipy.org/doc/numpy/reference/arrays.indexing.html
-        sshp = self.shape
-        for x in xrange(rank):
-            sl=slice_[x]
-            if isinstance(sl, int):
-                if sl < 0:
-                    slice_[x] = sshp[x] + slice_[x]
-            elif isinstance(sl, list):
-                for i in xrange(len(sl)):
-                    if sl[i] < 0:
-                        sl[i] = sshp[x] + sl[i]
-                slice_[x] = sl
-            elif isinstance(sl, slice):
-                sl_ = [sl.start, sl.stop, sl.step]
-                if sl_[0] is not None and sl_[0] < 0:
-                    sl_[0] = sshp[x] + sl_[0]
-
-                if sl_[1] is not None and sl_[1] < 0:
-                    sl_[1] = sshp[x] + sl_[1]
-
-                if sl_[2] is not None and sl_[2] < 0:
-                    slice_[x] = slice(sl_[1],sl_[0],-sl_[2])
-                else:
-                    slice_[x] = slice(*sl_)
-
-        # Finally, make it a tuple
-        return tuple(slice_)
 
     @property
     def shape(self):
@@ -170,12 +117,12 @@ class NumericValue(AbstractSimplexParameterValue):
         self._storage.expand(self.shape, 0, self.shape[0])
 
     def __getitem__(self, slice_):
-        slice_ = self._fix_slice(slice_)
+        slice_ = fix_slice(slice_, self.shape)
 
         return self._storage[slice_]
 
     def __setitem__(self, slice_, value):
-        slice_ = self._fix_slice(slice_)
+        slice_ = fix_slice(slice_, self.shape)
 
         self._storage[slice_] = value
 
@@ -205,7 +152,7 @@ class FunctionValue(AbstractComplexParameterValue):
         pass
 
     def __getitem__(self, slice_):
-        slice_ = self._fix_slice(slice_)
+        slice_ = fix_slice(slice_, self.shape)
 
         total_indices = prod(self.shape)
         x = np.arange(total_indices).reshape(self.shape)[slice_] # CBM TODO: This is INDEX based evaluation!!!
@@ -214,7 +161,7 @@ class FunctionValue(AbstractComplexParameterValue):
 
     def __setitem__(self, slice_, value):
         if is_well_formed_where(value):
-            slice_ = self._fix_slice(slice_)
+            slice_ = fix_slice(slice_, self.shape)
             if len(slice_) == 1 and isinstance(slice_[0], int) and slice_ < len(self._storage[0]):
                 self._storage[0][slice_[0]] = value
             else:
@@ -271,14 +218,14 @@ class RecordValue(AbstractComplexParameterValue):
         self._storage.expand(self.shape, 0, self.shape[0])
 
     def __getitem__(self, slice_):
-        slice_ = self._fix_slice(slice_)
+        slice_ = fix_slice(slice_, self.shape)
 
         ret = self._storage[slice_]
 
         return ret
 
     def __setitem__(self, slice_, value):
-        slice_ = self._fix_slice(slice_)
+        slice_ = fix_slice(slice_, self.shape)
 
         self._storage[slice_] = value
 
@@ -304,13 +251,13 @@ class ArrayValue(AbstractComplexParameterValue):
         self._storage.expand(self.shape, 0, self.shape[0])
 
     def __getitem__(self, slice_):
-        slice_ = self._fix_slice(slice_)
+        slice_ = fix_slice(slice_, self.shape)
 
         ret = self._storage[slice_]
 
         return ret
 
     def __setitem__(self, slice_, value):
-        slice_ = self._fix_slice(slice_)
+        slice_ = fix_slice(slice_, self.shape)
 
         self._storage[slice_] = value
